requirements.txt
fastapi
uvicorn
requests
pydantic
prometheus-client

# RAG
chromadb
sentence-transformers

# Utils
python-dotenv
pyyaml

